requirements.txt
deemix
deezer-py
spotipy
mutagen
python-dotenv
pyyaml
diskcache
textual